added reminderstart & stop button to homepage for testing + created /test  reactroute app.js for home page
