+++ my-react-app/src/Pages/Home.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const setReminders = async () => {
+const startReminders = async () => {
   try {
     let options = {
       method: "POST",
@@ -16,6 +16,22 @@ const setReminders = async () => {
   } 
 };
 
+const stopReminders = async () => {
+  try {
+    let options = {
+      method: "POST",
+      headers: {
+      "Content-Type": "application/json"},
+    }
+    let results = await fetch("/reminders-stop", options)
+    let notification = await results.json();
+    console.log(notification.message)
+    }
+    catch (error) {
+    console.log(error)
+  } 
+};
+
 
 function Home() {
   return (
@@ -23,7 +39,8 @@ function Home() {
 
       <p> Home Page </p>
       
-      <button onClick={setReminders}> Reminders </button>
+      <button onClick={startReminders}> Start Reminders </button>
+      <button onClick={stopReminders}> Stop Reminders </button>
 
 
     </div>
